+++ src/main/java/com/odeene/formacion/soap/model/City.java
@@ -2,7 +2,7 @@
 // Este archivo ha sido generado por la arquitectura JavaTM para la implantación de la referencia de enlace (JAXB) XML v2.3.2 
 // Visite <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Todas las modificaciones realizadas en este archivo se perderán si se vuelve a compilar el esquema de origen. 
-// Generado el: 2020.11.13 a las 09:52:17 AM CET 
+// Generado el: 2020.11.19 a las 05:39:55 PM CET 
 //
 
 
@@ -24,7 +24,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="cityId" type="{http://www.w3.org/2001/XMLSchema}int"/&gt;
+ *         &lt;element name="cityId" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
  *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
  *         &lt;element name="population" type="{http://www.w3.org/2001/XMLSchema}int"/&gt;
  *       &lt;/sequence&gt;
@@ -43,7 +43,7 @@ import javax.xml.bind.annotation.XmlType;
 })
 public class City {
 
-    protected int cityId;
+    protected long cityId;
     @XmlElement(required = true)
     protected String name;
     protected int population;
@@ -52,7 +52,7 @@ public class City {
      * Obtiene el valor de la propiedad cityId.
      * 
      */
-    public int getCityId() {
+    public long getCityId() {
         return cityId;
     }
 
@@ -60,7 +60,7 @@ public class City {
      * Define el valor de la propiedad cityId.
      * 
      */
-    public void setCityId(int value) {
+    public void setCityId(long value) {
         this.cityId = value;
     }
