+++ src/main/java/com/odeene/formacion/soap/model/Province.java
@@ -2,7 +2,7 @@
 // Este archivo ha sido generado por la arquitectura JavaTM para la implantación de la referencia de enlace (JAXB) XML v2.3.2 
 // Visite <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Todas las modificaciones realizadas en este archivo se perderán si se vuelve a compilar el esquema de origen. 
-// Generado el: 2020.11.13 a las 09:52:17 AM CET 
+// Generado el: 2020.11.19 a las 05:39:55 PM CET 
 //
 
 
@@ -24,7 +24,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="provinceId" type="{http://www.w3.org/2001/XMLSchema}int"/&gt;
+ *         &lt;element name="provinceId" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
  *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
@@ -41,7 +41,7 @@ import javax.xml.bind.annotation.XmlType;
 })
 public class Province {
 
-    protected int provinceId;
+    protected long provinceId;
     @XmlElement(required = true)
     protected String name;
 
@@ -49,7 +49,7 @@ public class Province {
      * Obtiene el valor de la propiedad provinceId.
      * 
      */
-    public int getProvinceId() {
+    public long getProvinceId() {
         return provinceId;
     }
 
@@ -57,7 +57,7 @@ public class Province {
      * Define el valor de la propiedad provinceId.
      * 
      */
-    public void setProvinceId(int value) {
+    public void setProvinceId(long value) {
         this.provinceId = value;
     }
